cleaner

+++ GoogleAuthTest/Models/YouTubeVideoMetricsRecord.cs
@@ -10,11 +10,11 @@ namespace GoogleAuthTest.Models
         public string VideoUrl { get; set; }
         public string AuthorDisplayName { get; set; }
         public int VideoDurationSeconds { get; set; }
-        public ulong Views { get; set; }
-        public ulong Likes { get; set; }
-        public ulong Dislikes { get; set; }
+        public int Views { get; set; }
+        public int Likes { get; set; }
+        public int Dislikes { get; set; }
         public int Shares { get; set; }
-        public ulong Replies { get; set; }
+        public int Replies { get; set; }
         public int EngagementTotal { get; set; }
         public DateTime PostDateTime { get; set; }
         public DateTime CreateDateTime { get; set; }
